# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Basic API",
        "description": "L5 Swagger API with Laravel basic",
        "contact": {
            "email": "nguyen.an.thoi.212@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/laravel_basic/an.thoi.laravel/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/status": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get List Status",
                "description": "Get List Status",
                "operationId": "GetListStatus",
                "responses": {
                    "200": {
                        "description": "Get List Status Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Add status",
                "description": "Add status",
                "operationId": "addStatus",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Add status",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "default": "Hello Everyone"
                                },
                                "content": {
                                    "type": "string",
                                    "default": "I am admin ahihihi...."
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edit profile successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/status/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get Status By Id",
                "description": "Get Status By Id",
                "operationId": "GetStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get status Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Edit Status By Id",
                "description": "Edit Status By Id",
                "operationId": "EditStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "edit status",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "default": "Ahihi..."
                                },
                                "content": {
                                    "type": "string",
                                    "default": "Still ahihihi...."
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edit status Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete Status By Id",
                "description": "Delete Status By Id",
                "operationId": "DeleteStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Edit status Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/{friend_list_id}/status": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get Friend Status",
                "description": "Get Friend Status",
                "operationId": "GetFriendStatus",
                "parameters": [
                    {
                        "name": "friend_list_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Friend Status Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/friend/{friend_list_id}/status/{status_id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get Friend Status By Id",
                "description": "Get Friend Status By Id",
                "operationId": "GetFriendStatusById",
                "parameters": [
                    {
                        "name": "friend_list_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "status_id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Friend Status By Id Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/new-feed": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get New Feed",
                "description": "Get New Feed",
                "operationId": "GetNewFeed",
                "responses": {
                    "200": {
                        "description": "Get New Feed Successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "A short description for security scheme",
            "name": "Authorization",
            "in": "header"
        }
    }
}